Break the eBPF program up
